requirements.txt
Django==3.2.9
django-ckeditor==6.5.1
django-environ==0.9.0
django-taggit==3.1.0
gunicorn==20.1.0
Pillow==9.4.0
requests==2.28.2
whitenoise==6.4.0
django-cloudinary-storage==0.3.0
celery==5.2.7
redis==4.5.1
dj-database-url==1.2.0
psycopg2-binary==2.9.5
